requirements.txt
pypng
opencv-python
opencv-contrib-python
wheel
scikit-learn
tensorflow==2.17.0
keras==3.5.0
matplotlib
prettytable
ptable
pyyaml
pandas